Use termion for input
Now instead of using the stdin library for input the shell uses termion in raw mode.

The code does the same as before for now but with termion, in which will be easier to add keybinds and special letter.

Deleted the prompt test because now it's useless and has to be done again.