Now shell uses rustyline instead of termion
After having some problems with termion and the fact that the shell doesn't the full terminal UI capabilities of something like tui, I found rustyline. Rustyline is a simple editable input crate which uses termion and makes the process easy, is simpler than termion or tui, but it's suitable for the project, for now.